Corregidos los scripts de las bases de datos. Corregido error en usuario y parametro desciption.

+++ app/parametro/parametro_description.php
@@ -159,7 +159,7 @@ $parametro_description =
                 ]
             ],
             'rango_hasta_parametro' => [
-                'type' => 'numeric',
+                'type' => 'integer',
                 'not_null' => [],
                 'rules' => [
                     'validations' => [
